+++ packages/client/src/rtc/Subscriber.ts
@@ -32,6 +32,19 @@ export class Subscriber {
   private isIceRestarting = false;
   private iceRestartTimeout?: NodeJS.Timeout;
 
+  // workaround for the lack of RTCPeerConnection.getConfiguration() method in react-native-webrtc
+  private _connectionConfiguration: RTCConfiguration | undefined;
+
+  /**
+   * Returns the current connection configuration.
+   *
+   * @internal
+   */
+  get connectionConfiguration() {
+    if (this.pc.getConfiguration) return this.pc.getConfiguration();
+    return this._connectionConfiguration;
+  }
+
   /**
    * Constructs a new `Subscriber` instance.
    *
@@ -77,6 +90,7 @@ export class Subscriber {
    */
   private createPeerConnection = (connectionConfig?: RTCConfiguration) => {
     const pc = new RTCPeerConnection(connectionConfig);
+    this._connectionConfiguration = connectionConfig;
     pc.addEventListener('icecandidate', this.onIceCandidate);
     pc.addEventListener('track', this.handleOnTrack);