+++ packages/react-native-sdk/expo-config-plugin/__tests__/withAndroidPermissions.test.ts
@@ -21,6 +21,7 @@ describe('withStreamVideoReactNativeSDKAndroidPermissions', () => {
         'android.permission.POST_NOTIFICATIONS',
         'android.permission.FOREGROUND_SERVICE',
         'android.permission.FOREGROUND_SERVICE_MICROPHONE',
+        'android.permission.FOREGROUND_SERVICE_MEDIA_PROJECTION',
         'android.permission.BLUETOOTH',
         'android.permission.BLUETOOTH_CONNECT',
         'android.permission.BLUETOOTH_ADMIN',