+++ sample-apps/react/egress-composite/src/hooks/useInitializeClient.ts
@@ -1,7 +1,7 @@
 import { useEffect, useMemo } from 'react';
 import {
-  StreamVideoClient,
   Call,
+  StreamVideoClient,
   StreamVideoParticipant,
 } from '@stream-io/video-react-sdk';
 
@@ -57,6 +57,10 @@ const useJoinCall = ({
 
   useEffect(() => {
     if (!client || !enabled) return;
+    
+    // the recorder system doesn't have any device attached
+    // call.camera.disable();
+    // call.microphone.disable();
 
     const callJoinPromise = call.join();
     return () => {