+++ sample-apps/react/react-dogfood/pages/join/[callId].tsx
@@ -3,6 +3,7 @@ import { useRouter } from 'next/router';
 import {
   Call,
   CallingState,
+  CallRequest,
   StreamCall,
   StreamVideo,
   StreamVideoClient,
@@ -43,6 +44,9 @@ const CallRoom = (props: ServerSideCredentialsProps) => {
   const callId = router.query['callId'] as string;
   const callType = (router.query['type'] as string) || 'default';
 
+  // support for connecting to any application using an API key and user token
+  const apiKeyOverride = !!router.query['api_key'];
+
   const isDemoEnvironment = useIsDemoEnvironment();
   useEffect(() => {
     if (!isDemoEnvironment) return;
@@ -60,11 +64,19 @@ const CallRoom = (props: ServerSideCredentialsProps) => {
     }
   }, [callId, isDemoEnvironment, router]);
 
-  const { user, gleapApiKey } = props;
+  const { apiKey, userToken, user, gleapApiKey } = props;
 
   const environment = useAppEnvironment();
   const fetchAuthDetails = useCallback(
     async (init?: RequestInit) => {
+      if (apiKeyOverride) {
+        return {
+          apiKey,
+          token: userToken,
+          userId: user.id || '!anon',
+        } satisfies CreateJwtTokenResponse;
+      }
+
       const params = {
         user_id: user.id || '!anon',
         environment,
@@ -75,7 +87,7 @@ const CallRoom = (props: ServerSideCredentialsProps) => {
         init,
       ).then((res) => res.json() as Promise<CreateJwtTokenResponse>);
     },
-    [environment, user.id],
+    [apiKey, apiKeyOverride, environment, user.id, userToken],
   );
 
   const tokenProvider = useCallback(
@@ -148,10 +160,18 @@ const CallRoom = (props: ServerSideCredentialsProps) => {
   }, [callId, callType, client]);
 
   useEffect(() => {
-    call?.getOrCreate().catch((err) => {
+    if (!call) return;
+    // "restricted" is a special call type that only allows
+    // `call_member` role to join the call
+    const data: CallRequest =
+      callType === 'restricted'
+        ? { members: [{ user_id: user.id || '!anon', role: 'call_member' }] }
+        : {};
+
+    call.getOrCreate({ data }).catch((err) => {
       console.error(`Failed to get or create call`, err);
     });
-  }, [call]);
+  }, [call, callType, user.id]);
 
   // apple-itunes-app meta-tag is used to open the app from the browser
   // we need to update the app-argument to the current URL so that the app